fix warnings.

+++ include/fast_float/bigint.h
@@ -71,9 +71,7 @@ template <uint8_t size> struct stackvec {
   }
 
   // set the length, without bounds checking.
-  FASTFLOAT_CONSTEXPR14 void set_len(uint8_t len) noexcept {
-    length = len;
-  }
+  FASTFLOAT_CONSTEXPR14 void set_len(uint8_t len) noexcept { length = len; }
 
   constexpr uint8_t len() const noexcept { return length; }
 
@@ -101,7 +99,7 @@ template <uint8_t size> struct stackvec {
   FASTFLOAT_CONSTEXPR20 void extend_unchecked(limb_span s) noexcept {
     limb *ptr = data + length;
     std::copy_n(s.ptr, s.len(), ptr);
-    set_len(len() + s.len());
+    set_len(uint8_t(len() + s.len()));
   }
 
   // try to add items to the vector, returning if items were added
@@ -304,7 +302,7 @@ FASTFLOAT_CONSTEXPR20 bool large_add_from(stackvec<size> &x, limb_span y,
   // the effective x buffer is from `xstart..x.len()`, so exit early
   // if we can't get that current range.
   if (x.len() < start || y.len() > x.len() - start) {
-    FASTFLOAT_TRY(x.try_resize(y.len() + start, 0));
+    FASTFLOAT_TRY(x.try_resize(uint8_t(y.len() + start), 0));
   }
 
   bool carry = false;
@@ -547,7 +545,7 @@ struct bigint : pow5_tables<> {
       limb *first = vec.data;
       limb *last = first + n;
       ::std::fill(first, last, 0);
-      vec.set_len(n + vec.len());
+      vec.set_len(uint8_t(n + vec.len()));
       return true;
     } else {
       return true;
@@ -584,8 +582,8 @@ struct bigint : pow5_tables<> {
 
   // get the number of bits in the bigint.
   FASTFLOAT_CONSTEXPR20 uint16_t bit_length() const noexcept {
-    uint16_t lz = ctlz();
-    return uint16_t(limb_bits * vec.len()) - lz;
+    uint8_t lz = ctlz();
+    return limb_bits * vec.len() - lz;
   }
 
   FASTFLOAT_CONSTEXPR20 bool mul(limb y) noexcept { return small_mul(vec, y); }
